id: forbid searching for branches created late
A subsequent commit will make the CLI IDs of committed files unambiguous
with both themselves and branch IDs, but this requires branch IDs to be
frozen at the point of `IdDb` creation (if not, the creation of a branch
could force an existing committed file ID to be changed). Therefore,
only consult the branches known at the point of `IdDb` creation when
searching for a branch. This also simplifies the code.

Ideally, this invariant (of the lifetime of `IdDb` lasting only as long
as the repo is unmodified) could be upheld in the Rust type system (e.g.
tying `IdDb` to the lifetime of a non-repo-modifying `Context`, and when
the repo needs to be modified, the `Context` must be exchanged for a new
repo-modifying context), but this seems like a significant undertaking,
so this has been left for later.