id: make `from_str` take `IdDb` (as self)
Now that construction of `IdDb` is more expensive (due to diffing
commits against their parents), it's more important that we reduce the
number of times it is constructed, so make `from_str` take `IdDb` (as
self) instead of constructing it every time `from_str` is called.

To avoid a (reasonable) Clippy finding, rename this function to
`parse_str`.