id: remove dependence on legacy code for commits
Also, simplified commit ID parsing in `IdDb`. It used to be done before
uncommitted files, committed files, and branches if the commit ID given
was at least 4 characters long (with another redundant parse after),
and after uncommitted files, committed files, and branches otherwise.
But it does not matter whether it's done before or after, since commit
IDs never collide with any of those (which all start with g-z), so do it
before in all cases.

In the absence of a way to unit test `id.rs`, I've added integration
tests once again. Once we migrate `id.rs` completely off legacy code,
we may be able to narrow down what exactly is needed (probably not the
whole `Context`) and thus be able to make unit tests.