id: unambiguous uncommitted file CLI IDs
They also do not collide with committed file IDs or branch IDs.

Sample output:

```
╭┄000 [Unassigned Changes]
┊   g0 A d
┊
┊╭┄ux [b00] (no commits)
├╯
┊
┊╭┄wo [work]
┊│   h0 A e
┊●   89f03ba test commit
┊│     j0 A a
┊│     k0 A b
┊│     l0 A c
├╯
┊
┊╭┄i0 [i0] (no commits)
├╯
┊
┴ dba337d (common base) [origin/main] 2025-11-11 Merge pull request #2254 from ralphmodales/credent
```

I had to do some contortions in `context_info()` to ensure that
`repo`, and hence the indirect borrowing of `ctx`, was dropped
before `but_hunk_assignment::assignments_with_fallback()`, which
needs to borrow `ctx` mutably, was called. This will be fixed once
`context_info()` does not take `ctx`, but only what it needs (most
likely, just the `RefInfo` and the hunk assignments).

There is also a change to an error message in a test
(`crates/but/tests/but/journey.rs`), likely due to uncommitted files
being read at `IdMap` construction time, instead of when it is used to
parse a string.

TODO need to write tests