UI fixes (#11468)
* Fix erroneous github popup

When switching from a github repo to a non-github repo you still get
the `UnassignedViewForgePrompt` because the _determinedForgeType didn't get updated.

* make commit details also a drag/drop target

* fix: remove unused animations and scale transform

Don’t scale the dropzone container using the `transform` property because if the dropzone is too large, it could become excessively large and be cropped by the parent container.

* lint and check fixes

---------

Co-authored-by: Pavel Laptev <pawellaptew@gmail.com>