Fix SSR related issues on marketing site
We ought to  take a more wholistic approach to SSR at some point because
I really just put a bandaid on the problem here.

I did some research into what a well structured devide between server 
side and client side rendering and there doesn’t seem to be a great 
idiom on the surface, but I’m certian there are some idioms that we 
could enforce with ESLint to make this much more pricipaled.