id: narrower constructor arguments (instead of `Context`)
Instead of taking `Context`, make `IdMap`'s constructor take
only the specific things needed:
- the stacks
- a function that can generate a list of files given a commit ID and its
  parent commit ID
- hunk assignments

Upon closer inspection, only the first is needed for a command that
operates on commits (and not on files); for performance reasons, these
commands should not need to diff commits, so construction of `IdMap` has
been split into two - one part only for commit and branch information,
and a second part for file information (to be filled immediately if
the command needs it, or when an error message of the form "42 is a
file ID, but only commit IDs are supported by this command" - in this
case, we need the file information to be able to generate such an error
message). In this commit, however, all commands still use both parts -
optimization of each command is left as later work.

Obtaining the necessary things from `Context` is non-trivial, and I
thought of putting the relevant code in `Context`, but it seemed better
to not crowd `Context`, so I put it in `IdMap` instead.