Centralize mutations of branchesSelection
In various places, fields like `inWorkspace` got dropped because of full object overwrites on branchesSelection. This behavior improves by doing updates on the object when possible, but not all cases are solved. When actually mutating a branch state (e.g. unapply), it may still show up in parts of the UI wrong. To fix this a bigger change is needed, where that state becomes properly derived and is no longer part of the selection object, I added a TODO for that.