id: test that IDs are case sensitive
The current behavior is that IDs are case sensitive. Enforce that in a
test to show that this is on purpose.

In the ID parsing code that evaluates if the ID is a commit, there
is an `is_ascii_hexdigit()` check which allows uppercase A-F, but the
code immediately following it does not permit uppercase A-F (it uses
`oid.to_string().starts_with()`, which only permits lowercase). Change
the `is_ascii_hexdigit()` check to make it clearer that only lowercase
a-f are permitted, even for commits.