Use `BString` as editor return value to not enforce encodings.
This will make it harder to call from legacy code, but will make it
clear for modern code that we don't know the encoding and shouldn't
lighthartedly butcher it by converting to unicode lossily.

Note that all callers now state explicitly that they handle the message
lossily, and we should fix this once the code is modernised.

While at it, also avoid rewriting line endings in favor of reusing
the ones that the user specified.