Replace Ubuntu 24.04 builds with Ubuntu 22.04
GitButler builds fine on Ubuntu 22.04, and building on this older system
should provide an overall wider range of compatibility with older
distros due to the older GLIBC version (2.35, as opposed to 2.38 on
Ubuntu 24.04).

There are known forward-compatibility issues with the AppImage built on
22.04, just like there are known backward-compatibility issues with the
AppImage built on Ubuntu 24.04. As we're demoting the AppImage to
experimental anyway, this is acceptable.