id: deduplicate uncommitted files
The input to `IdMap`'s constructor may contain multiple
`HunkAssignments` for one uncommitted file, so teach it to deduplicate
them. Committed files do not have the same problem, as can be seen from
the included test (but the test tests committed files anyway to prevent
regressions).

The test uses `status`, which is not ideal (this is an `IdMap` concern),
but I have included a TODO describing how and when it can be removed.