id, rub: uncommitted hunk to stack
This is the beginning of hunk CLI IDs. As of this commit, the user
can use a CLI ID to assign an uncommitted hunk to a stack if they know
what the hunk's ID is (right now, there is no way other than knowing
GitButler internals; a "diff" or "show" command that presents hunk CLI
IDs to the user is left for future work).

Note that `uncommitted_and_committed_file_cli_ids` in
crates/but/tests/but/command/status.rs should not be deleted yet, since
we don't have tests for committed hunk CLI IDs yet.