Perform worktree checkout first in V3 apply (#11553)
It's most likely to fail and the most complex part of the checkout,
so it should fail as early as possible without any other data being
written already.

Currently this leaves the worktree in an invalid/partial state, but
it's something that can be improved on once it's done by Gitoxide.