id: ensure that branch IDs work
When provided a branch ID, `IdMap` checks all branches to see if the
given ID is a substring of the branch name. This works with most keys,
since `IdMap` attempts as much as possible to generate a key that is a
substring of the branch name. But sometimes that is not possible, and in
that case, the substring search is not enough.

Teach `IdMap` to record when the generated key is not a substring of the
branch name, and to use that record when resolving an ID.