Fix reword in certain cases
The issue was the order in which we created the graph. The traversal 
that we did could get to a merge commit without first visiting all of 
it’s parents which was causing that missing parent issue.

By first creating all the nodes and then all the parents, we are able to
avoid this issue
a