Improvements to rebase engine for better composibility
This commit introduces a “editor in, successful rebase out” pattern.

The editor in theory should be enough to implement a given manipulation 
like “insert_blank_commit” or “reword_commit”.

If I want to compose “insert_bank_commit” and “reword_commit”, I don’t 
want insert_blank_commit to actually materialize it’s changes yet - but,
it probably should error if the insertion failed.

If we are in a world where both of these functions take in an editor and
return references to their added options with selectors, we can 
reasonabily achieve this composibility.

If “insert_blank_commit” only calls `.rebase()?`, it will return an Err,
if the insertion failed in memory, and then as a caller, I can then 
decide to materialize that result, OR, I can choose to convert that 
rebase result into another editor for “reword_commit” to use - which 
provides us with some really cool composible primateives