Update rebase engine to be more composable.
This switches to using a new editor in, rebase outcome out.

The rebase outcome can either be turned into a new editor for further
operations being performed in-memory, or can be materialized.

With this pattern, it also estabilishes the idea of returning selectors rather
than direct object ids.

This is because object ids from the initial rebase can change after
materializing (if we have things like delayed signing implemented) - so we
return a selector that can either be given to subsequent operations, or have
it's target fetched after materialization for use in a given frontend.