id: hunk assignments in uncommitted file `CliId`
This prevents double work, and prepares for future work that will unify
`CliId`s for uncommitted files and hunks.

There is some code in `rub` that generates every uncommitted file
`CliId` while parsing ranges, which will incur additional memory
copying, but this will hopefully go away when range parsing is
refactored to be in `IdMap` instead (in the future).