id: cleanup `CliId` traits
`CliId` does not need `Hash`, so remove it. Before some recent
significant changes, `Ord` was more crucial because we needed to
sort IDs when the user provided an ambiguous entity. However, most
ambiguities have been removed, so we no longer need to sort. The only
ambiguities remaining are:
- branch vs branch (handled by switching `branch_name_to_cli_id` to a
  `BTreeMap` to make the order consistent)
- commit ID vs unassigned area (handled by checking commits first)

So `Ord` is no longer necessary, so remove it too. Sorting is necessary
for tests because we list all IDs there, so add that functionality to
the tests.

As for `PartialEq`, change it to only check what's needed.

All this prepares `CliId` for a subsequent commit in which we add
a field that cannot have these traits derived easily.