id: remove resolve_uncommitted_file_or_unassigned
This also removes some complicated parts of the code, including
`UncommittedFile` needing to be indexable in a `BTreeSet` both by (stack
ID, path) and by CLI ID (as of this commit, it only needs to be stored
in a regular map). This also makes it possible to use a `NonEmpty`
in `CliId::UncommittedFile`, with associated code improvements in its
users, because we no longer need a fallback CLI ID if the uncommitted
file does not have an associated ID (this situation is now impossible).

This commit also eliminates some duplicate work in callers. Previously,
some callers obtained hunk assignments from elsewhere and reconciled
them with `IdMap` using `resolve_uncommitted_file_or_unassigned`.
These callers have been changed to use the data structures in `IdMap`
directly.