Prevent dropzone-label-icon from animating when hidden (#11594)
This icon was always rendered and animated for any dropzone with a
label, and the lanes in the workspace have dropzones with labels.
Thus, any non-empty workspace would suffer constant animations.

WebkitGTK (i.e. Linux) has terrible animation performance, but even
running in Chrome the CPU usage was noticable for me.

By only animating the icon when it's actually visible, we save a lot of
CPU cycles.