Remove `gitbutler-testsupport` from `but-hunk-dependencies`
This makes it testable again as it won't rebuild the world, twice,
each time tests are run.

IMPORTANT NOTE: It looks like the `merge-commit` fixture, albeit the same,
now shows to me legitimate extra hunks that weren't there before.
Previously, the line 5 hunk was 'missed', but it's unclear why.
Now it seems better than before, but in any case, it's a testing issue
if it is an issue, as the implementation didn't change.

Maybe with future improvements, expectations will change as well.