move `snapshot` command under `oplog` family
This moves the `but snapshot` command to `but oplog snapshot` and adds a
way to filter the oplog list to only snapshots.

It also removes use of the pager from the normal `but oplog` listing so
we can actually restore stuff (unless `—all` is supplied).

The `—all` flag is not ideal - it just sends `usize::MAX` to the api. 
It would probably be better to have something that is listing out the
data as it walks the Git parentage in real time, rather than cache all
the data and send it as one big result, but we’ll have to push that to
a later implementation.